Bug fixes, code improvements
Fixed a few bugs regarding the semantically correct, expected functionality of a couple of global regular and template functions. There was also a couple of minor semantical bugs (unexpected behavior for certain textual inputs that do not contain specified needle argument (char, const char*, StringType). I accidentally stumbled upon this during the testing phase of the generic templatized version of split with a character string input that didn't contain the specified needle value.
Most of the functions work correctly, but some of them still have to be tested, thoroughly.