couple of small updates, bug fixes
Fixed a couple of small bugs regarding the way the 'split' global template and regular functions work. Now they are semantically more correct in the sense that they now return a non-empty vector with at least 1 string element even if specified needle is not found in source string. For example, even if specified needle is not found in source string or if specified needle's length is greater than or equal to specified source string's length, the function now returns a non-empty vector with specified source string (src) as its only element.